move markdown-it setup into markdown-converter util
